news/api/v1/serializers.py
from rest_framework import serializers
from news.models import Category, Post


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ['id', 'title', 'slug']


class PostSerializer(serializers.ModelSerializer):
    category = CategorySerializer()

    class Meta:
        model = Post
        fields = ['id', 'category', 'image', 'title', 'body', 'created', 'updated', 'views_count']
